test: Generate test annotations rather than doing them dynamically
It is expensive to perform 100 checks against 1k names on every
startup. Instead, move these to generated files that can be reviewed
out of bounds on change. Add them to generated-bindata for verification
and regen.
