+++ test/extended/include.go
@@ -1,5 +1,7 @@
 package extended
 
+//go:generate go run ./util/annotate -- ./util/annotate/generated/zz_generated.annotations.go
+
 import (
 	_ "k8s.io/kubernetes/test/e2e"
