+++ test/extended/util/annotate/OWNERS
@@ -1,0 +1,5 @@
+reviewers:
+  - smarterclayton
+approvers:
+  - smarterclayton
+  - derekwaynecarr
